chore: split middleware files into individual files per middleware
