+++ internal/api/middleware_recovery.go
@@ -1,0 +1,59 @@
+package api
+
+import (
+	"log/slog"
+	"net/http"
+)
+
+// RecoveryMiddleware catches panics and returns 500.
+func RecoveryMiddleware(next http.Handler) http.Handler {
+	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+		rw := &recoveryResponseWriter{ResponseWriter: w}
+		defer func() {
+			if err := recover(); err != nil {
+				if err == http.ErrAbortHandler {
+					panic(err)
+				}
+				slog.Error("panic recovered", "err", err, "path", r.URL.Path)
+				if rw.wroteHeader {
+					panic(http.ErrAbortHandler)
+				}
+				writeError(rw, http.StatusInternalServerError, "internal server error")
+			}
+		}()
+		next.ServeHTTP(rw, r)
+	})
+}
+
+type recoveryResponseWriter struct {
+	http.ResponseWriter
+	wroteHeader bool
+}
+
+func (rw *recoveryResponseWriter) WriteHeader(code int) {
+	if rw.wroteHeader {
+		return
+	}
+	rw.wroteHeader = true
+	rw.ResponseWriter.WriteHeader(code)
+}
+
+func (rw *recoveryResponseWriter) Write(data []byte) (int, error) {
+	if !rw.wroteHeader {
+		rw.wroteHeader = true
+	}
+	return rw.ResponseWriter.Write(data)
+}
+
+func (rw *recoveryResponseWriter) Flush() {
+	if !rw.wroteHeader {
+		rw.wroteHeader = true
+	}
+	if f, ok := rw.ResponseWriter.(http.Flusher); ok {
+		f.Flush()
+	}
+}
+
+func (rw *recoveryResponseWriter) Unwrap() http.ResponseWriter {
+	return rw.ResponseWriter
+}
