+++ internal/runner/middleware_auth.go
@@ -1,0 +1,21 @@
+package runner
+
+import "net/http"
+
+// AuthMiddleware checks for valid API keys.
+func AuthMiddleware(apiKeys map[string]struct{}) func(http.Handler) http.Handler {
+	return func(next http.Handler) http.Handler {
+		return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+			apiKey := r.Header.Get("X-Api-Key")
+			if apiKey == "" {
+				writeError(w, http.StatusUnauthorized, "missing API key")
+				return
+			}
+			if _, ok := apiKeys[apiKey]; !ok {
+				writeError(w, http.StatusUnauthorized, "invalid API key")
+				return
+			}
+			next.ServeHTTP(w, r)
+		})
+	}
+}
