+++ internal/runner/middleware_logging.go
@@ -1,0 +1,24 @@
+package runner
+
+import (
+	"log/slog"
+	"net/http"
+	"time"
+)
+
+// LoggingMiddleware logs HTTP requests.
+func LoggingMiddleware(next http.Handler) http.Handler {
+	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+		if r.URL.Path == "/healthz" {
+			next.ServeHTTP(w, r)
+			return
+		}
+		start := time.Now()
+		next.ServeHTTP(w, r)
+		slog.Info("request",
+			"method", r.Method,
+			"path", r.URL.Path,
+			"duration", time.Since(start),
+		)
+	})
+}
